backend/requirements.txt
fastapi==0.115.0
uvicorn==0.23.2
gunicorn==23.0.0
openai==1.12.0
psycopg2-binary==2.9.9
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.1
google-auth==2.38.0
PyJWT==2.10.1
numpy==1.26.4
python-dotenv